confusion matrix is int

+++ train.py
@@ -45,6 +45,6 @@ comet_logger.experiment.log_metrics(crown_metrics)
 
 #Confusion matrix
 comet_logger.experiment.log_confusion_matrix(
-    results.true_label,
-    results.label,
+    results.true_label.astype(np.int),
+    results.label.astype(np.int),
     labels=list(data_module.species_label_dict.keys()))
